pyproject.toml
[project]
name = "secure-agent-flow"
version = "0.1.0"
description = "A CrewAI-based secure agent flow for role analysis and policy creation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.186.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
